requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
bcrypt==4.0.1
billiard==3.6.4.0
bleach==5.0.1
blinker==1.5
Brotli==1.0.9
cachelib==0.9.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.7.0
dnspython==2.2.1
email-validator==1.3.0
Flask==2.1.3
Flask-BabelEx==0.9.4
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mailman==0.3.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.38.0
greenlet==2.0.1
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.1
kombu==5.2.4
MarkupSafe==2.1.1
mkdocs-material-extensions==1.1.1
numpy==1.23.5
packaging==21.3
pandas==1.5.2
passlib==1.7.4
Pillow==9.3.0
prompt-toolkit==3.0.33
pycparser==2.21
pydyf==0.5.0
pyparsing==3.0.9
pyphen==0.13.2
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
redis==4.3.5
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
tinycss2==1.2.1
vine==5.0.0
wcwidth==0.2.5
weasyprint==57.1
webencodings==0.5.1
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.10.0
zopfli==0.2.2
